@@ -7,21 +7,40 @@
 	<groupId>org.example</groupId>
 	<artifactId>java-basic-enhance</artifactId>
 	<version>1.0-SNAPSHOT</version>
+
+	<properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<java.version>11</java.version>
+		<spring.version>5.2.2.RELEASE</spring.version>
+  	</properties>
+
 	<build>
 		<plugins>
+			<!-- 指定jdk -->
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.8.1</version>
 				<configuration>
-					<source>8</source>
-					<target>8</target>
+					<skip>true</skip>
+					<source>${java.version}</source>
+					<target>${java.version}</target>
+					<encoding>${project.build.sourceEncoding}</encoding>
+				</configuration>
+			</plugin>
+			<!-- 跳过测试 -->
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-surefire-plugin</artifactId>
+				<version>3.0.0-M4</version>
+				<configuration>
+					<skip>true</skip>
+					<skipTests>true</skipTests>
 				</configuration>
 			</plugin>
 		</plugins>
-	</build>
-	<properties>
-		<spring.version>5.2.2.RELEASE</spring.version>
-	</properties>
+    </build>
+
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework</groupId>
@@ -31,7 +50,7 @@
 		<dependency>
 			<groupId>com.google.guava</groupId>
 			<artifactId>guava</artifactId>
-			<version>19.0</version>
+			<version>30.1-jre</version>
 		</dependency>
 	</dependencies>
 
